Added page routing + new logic for ending cards
Main page,
End page
Main page is our standard page we have been using with the searchbar and cards.

the "nextpage" property of the card object is now a string. To create and "end card" add an "e" before the number on the next page and increment the number. It needs to be unique.

Added the endpage link property and the min price.